requirements.txt
aiohttp==3.9.1
email_validator==2.1.0.post1
Flask==3.0.2
gpt4all==2.3.2
mysql_connector_repackaged==0.3.1
openai==1.16.1
psutil==5.9.8
PyQt6==6.6.1
PyQt6_sip==13.6.0
PySide6==6.6.2
PySide6==6.6.3.1
PySide6_Addons==6.6.2
PySide6_Essentials==6.6.2
Requests==2.31.0
Unidecode==1.3.8
